Both login and change schedule work partially now
The login works really well, but whenever  I go back to the dashboard after I try to change my schedule (which it doesn't appear to work in the front end but it does work in the backend), I end up logged off. Also, the tasks get a bit mixed up in the UI but they seem fine when I look at them in the database.